Refactoring CSharpServer class
