+++ .gitignore
@@ -12,3 +12,12 @@
 /C_Sharp_Server/Lynxmotion
 /C_Sharp_Server/TestLynxmotion
 /C_Sharp_Server/TestFunction
+/CSharpServer/CSharpClient/bin/Debug
+/CSharpServer/CSharpClient/obj
+/CSharpServer/CSharpServer/bin/Debug/netcoreapp2.0
+/CSharpServer/CSharpServer/auth.json
+/CSharpServer/Lynxmotion
+/CSharpServer/packages
+/CSharpServer/TestFunction
+/CSharpServer/TestLynxmotion
+/CSharpServer/.vs/CSharpServer/v15/Server/sqlite3
