+++ CSharpServer/CSharpClient/Properties/AssemblyInfo.cs
