--- C_Sharp_Server/C_Sharp_Server/auth.json
@@ -1,5 +1,0 @@
-{
-    "user": "wsurop",
-    "pass": "wsurop",
-    "dns": "35.177.37.56"
-}
